# swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0",
    "title": "Wayfarer",
    "description": "WayFarer is a public bus transportation booking server and you are looking at it's documentation.",
    "contact": {}
  },
  "host": "localhost:4000",
  "basePath": "/api/v1",
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/auth/signup": {
      "post": {
        "description": "This route enables new users to sign up",
        "summary": "/api/v1/auth/signup",
        "tags": [
          "Misc"
        ],
        "operationId": "AuthSignupPost",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Content-Type",
            "in": "header",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "Body",
            "in": "body",
            "required": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/~1api~1v1~1auth~1signuprequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      }
    },
    "/auth/signin": {
      "post": {
        "description": "Enables users to signin",
        "summary": "/api/v1/auth/signin",
        "tags": [
          "Misc"
        ],
        "operationId": "AuthSigninPost",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Content-Type",
            "in": "header",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "Body",
            "in": "body",
            "required": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/~1api~1v1~1auth~1signinrequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      }
    },
    "/trips": {
      "post": {
        "description": "Enable Admin to post a trip",
        "summary": "/api/v1/trips",
        "tags": [
          "Misc"
        ],
        "operationId": "TripsPost",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Content-Type",
            "in": "header",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "Body",
            "in": "body",
            "required": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/~1api~1v1~1tripsrequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      },
      "get": {
        "description": "Enables Users and Admins to get a trips",
        "summary": "/api/v1/trips",
        "tags": [
          "Misc"
        ],
        "operationId": "TripsGet",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Content-Type",
            "in": "header",
            "required": true,
            "type": "string",
            "description": ""
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      }
    },
    "/bookings": {
      "post": {
        "description": "Enable Users book a seat",
        "summary": "/api/v1/bookings",
        "tags": [
          "Misc"
        ],
        "operationId": "BookingsPost",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Content-Type",
            "in": "header",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "Body",
            "in": "body",
            "required": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/~1api~1v1~1bookingsrequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      }
    }
  },
  "definitions": {
    "/api/v1/auth/signuprequest": {
      "title": "/api/v1/auth/signuprequest",
      "example": {
        "email": "evuazeze.emmanuel@gmail.com",
        "password": "1234",
        "first_name": "Emmanuel",
        "last_name": "Evuazeze",
        "is_admin": true
      },
      "type": "object",
      "properties": {
        "email": {
          "type": "string"
        },
        "password": {
          "type": "string"
        },
        "first_name": {
          "type": "string"
        },
        "last_name": {
          "type": "string"
        },
        "is_admin": {
          "type": "boolean"
        }
      },
      "required": [
        "email",
        "password",
        "first_name",
        "last_name",
        "is_admin"
      ]
    },
    "/api/v1/auth/signinrequest": {
      "title": "/api/v1/auth/signinrequest",
      "example": {
        "email": "evuazeze.emmanuel@gmail.com",
        "password": "1234"
      },
      "type": "object",
      "properties": {
        "email": {
          "type": "string"
        },
        "password": {
          "type": "string"
        }
      },
      "required": [
        "email",
        "password"
      ]
    },
    "/api/v1/tripsrequest": {
      "title": "/api/v1/tripsrequest",
      "example": {
        "token": "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJzdWIiOjI5fQ.Bo3bgQh5QDjEJ_135QJftjBLK9tEAeEdV4FDYaYo-FU",
        "user_id": 29,
        "is_admin": true,
        "data": {
          "bus_id": 1,
          "origin": "Delta",
          "destination": "USA",
          "trip_date": "2019-07-30",
          "fare": 3500,
          "status": 1
        }
      },
      "type": "object",
      "properties": {
        "token": {
          "type": "string"
        },
        "user_id": {
          "type": "integer",
          "format": "int32"
        },
        "is_admin": {
          "type": "boolean"
        },
        "data": {
          "$ref": "#/definitions/Data"
        }
      },
      "required": [
        "token",
        "user_id",
        "is_admin",
        "data"
      ]
    },
    "Data": {
      "title": "Data",
      "example": {
        "bus_id": 1,
        "origin": "Delta",
        "destination": "USA",
        "trip_date": "2019-07-30",
        "fare": 3500,
        "status": 1
      },
      "type": "object",
      "properties": {
        "bus_id": {
          "type": "integer",
          "format": "int32"
        },
        "origin": {
          "type": "string"
        },
        "destination": {
          "type": "string"
        },
        "trip_date": {
          "type": "string"
        },
        "fare": {
          "type": "integer",
          "format": "int32"
        },
        "status": {
          "type": "integer",
          "format": "int32"
        }
      },
      "required": [
        "bus_id",
        "origin",
        "destination",
        "trip_date",
        "fare",
        "status"
      ]
    },
    "/api/v1/bookingsrequest": {
      "title": "/api/v1/bookingsrequest",
      "example": {
        "token": "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJzdWIiOjI5fQ.Bo3bgQh5QDjEJ_135QJftjBLK9tEAeEdV4FDYaYo-FU",
        "user_id": 29,
        "is_admin": false,
        "trip_id": 50,
        "bus_id": 9,
        "seat_number": 1
      },
      "type": "object",
      "properties": {
        "token": {
          "type": "string"
        },
        "user_id": {
          "type": "integer",
          "format": "int32"
        },
        "is_admin": {
          "type": "boolean"
        },
        "trip_id": {
          "type": "integer",
          "format": "int32"
        },
        "bus_id": {
          "type": "integer",
          "format": "int32"
        },
        "seat_number": {
          "type": "integer",
          "format": "int32"
        }
      },
      "required": [
        "token",
        "user_id",
        "is_admin",
        "trip_id",
        "bus_id",
        "seat_number"
      ]
    }
  },
  "tags": [
    {
      "name": "Misc",
      "description": ""
    }
  ]
}